+++ pkg/coderefs/coderefs_test.go
@@ -1,4 +1,4 @@
-package parse
+package coderefs
 
 import (
 	"regexp"
